+++ pkg/alertmanager/amcfg.go
@@ -27,6 +27,7 @@ import (
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/pkg/errors"
+	"github.com/prometheus-operator/prometheus-operator/pkg/alertmanager/validation"
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	monitoringv1alpha1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1alpha1"
 	"github.com/prometheus-operator/prometheus-operator/pkg/assets"
@@ -318,7 +319,7 @@ func (cb *configBuilder) getValidURLFromSecret(ctx context.Context, namespace st
 	}
 
 	url = strings.TrimSpace(url)
-	if _, err := ValidateURL(url); err != nil {
+	if _, err := validation.ValidateURL(url); err != nil {
 		return url, errors.Wrapf(err, "invalid URL %q in key %q from secret %q", url, selector.Key, selector.Name)
 	}
 	return url, nil
@@ -495,6 +496,18 @@ func (cb *configBuilder) convertReceiver(ctx context.Context, in *monitoringv1al
 		}
 	}
 
+	var telegramConfigs []*telegramConfig
+	if l := len(in.TelegramConfigs); l > 0 {
+		telegramConfigs = make([]*telegramConfig, l)
+		for i := range in.TelegramConfigs {
+			receiver, err := cb.convertTelegramConfig(ctx, in.TelegramConfigs[i], crKey)
+			if err != nil {
+				return nil, errors.Wrapf(err, "TelegramConfig[%d]", i)
+			}
+			telegramConfigs[i] = receiver
+		}
+	}
+
 	return &receiver{
 		Name:             makeNamespacedString(in.Name, crKey),
 		OpsgenieConfigs:  opsgenieConfigs,
@@ -506,6 +519,7 @@ func (cb *configBuilder) convertReceiver(ctx context.Context, in *monitoringv1al
 		VictorOpsConfigs: victorOpsConfigs,
 		PushoverConfigs:  pushoverConfigs,
 		SNSConfigs:       snsConfigs,
+		TelegramConfigs:  telegramConfigs,
 	}, nil
 }
 
@@ -521,7 +535,7 @@ func (cb *configBuilder) convertWebhookConfig(ctx context.Context, in monitoring
 		}
 		out.URL = url
 	} else if in.URL != nil {
-		url, err := ValidateURL(*in.URL)
+		url, err := validation.ValidateURL(*in.URL)
 		if err != nil {
 			return nil, err
 		}
@@ -577,7 +591,7 @@ func (cb *configBuilder) convertSlackConfig(ctx context.Context, in monitoringv1
 	if l := len(in.Actions); l > 0 {
 		actions = make([]slackAction, l)
 		for i, a := range in.Actions {
-			var action slackAction = slackAction{
+			action := slackAction{
 				Type:  a.Type,
 				Text:  a.Text,
 				URL:   a.URL,
@@ -602,9 +616,9 @@ func (cb *configBuilder) convertSlackConfig(ctx context.Context, in monitoringv1
 	}
 
 	if l := len(in.Fields); l > 0 {
-		var fields []slackField = make([]slackField, l)
+		fields := make([]slackField, l)
 		for i, f := range in.Fields {
-			var field slackField = slackField{
+			field := slackField{
 				Title: f.Title,
 				Value: f.Value,
 			}
@@ -738,7 +752,7 @@ func (cb *configBuilder) convertOpsgenieConfig(ctx context.Context, in monitorin
 	if l := len(in.Responders); l > 0 {
 		responders = make([]opsgenieResponder, 0, l)
 		for _, r := range in.Responders {
-			var responder opsgenieResponder = opsgenieResponder{
+			responder := opsgenieResponder{
 				ID:       r.ID,
 				Name:     r.Name,
 				Username: r.Username,
@@ -943,6 +957,38 @@ func (cb *configBuilder) convertPushoverConfig(ctx context.Context, in monitorin
 			return nil, err
 		}
 		out.HTTPConfig = httpConfig
+	}
+
+	return out, nil
+}
+
+func (cb *configBuilder) convertTelegramConfig(ctx context.Context, in monitoringv1alpha1.TelegramConfig, crKey types.NamespacedName) (*telegramConfig, error) {
+	out := &telegramConfig{
+		VSendResolved:        in.SendResolved,
+		APIUrl:               in.APIURL,
+		ChatID:               in.ChatID,
+		Message:              in.Message,
+		DisableNotifications: false,
+		ParseMode:            in.ParseMode,
+	}
+
+	if in.HTTPConfig != nil {
+		httpConfig, err := cb.convertHTTPConfig(ctx, *in.HTTPConfig, crKey)
+		if err != nil {
+			return nil, err
+		}
+		out.HTTPConfig = httpConfig
+	}
+
+	if in.BotToken != nil {
+		botToken, err := cb.store.GetSecretKey(ctx, crKey.Namespace, *in.BotToken)
+		if err != nil {
+			return nil, errors.Wrap(err, "failed to get bot token")
+		}
+		if botToken == "" {
+			return nil, fmt.Errorf("mandatory field %q is empty", "botToken")
+		}
+		out.BotToken = botToken
 	}
 
 	return out, nil
@@ -1379,6 +1425,12 @@ func (r *receiver) sanitize(amVersion semver.Version, logger log.Logger) error {
 		}
 	}
 
+	for _, conf := range r.TelegramConfigs {
+		if err := conf.sanitize(amVersion, withLogger); err != nil {
+			return err
+		}
+	}
+
 	return nil
 }
 
@@ -1405,6 +1457,11 @@ func (ogc *opsgenieConfig) sanitize(amVersion semver.Version, logger log.Logger)
 		level.Warn(logger).Log("msg", msg, "current_version", amVersion.String())
 		ogc.UpdateAlerts = nil
 	}
+	for _, responder := range ogc.Responders {
+		if err := responder.sanitize(amVersion, logger); err != nil {
+			return err
+		}
+	}
 
 	if ogc.APIKey != "" {
 		level.Warn(logger).Log("msg", "'api_key' and 'api_key_file' are mutually exclusive for OpsGenie receiver config - 'api_key' has taken precedence")
@@ -1421,6 +1478,13 @@ func (ogc *opsgenieConfig) sanitize(amVersion semver.Version, logger log.Logger)
 		ogc.APIKeyFile = ""
 	}
 
+	return nil
+}
+
+func (ops *opsgenieResponder) sanitize(amVersion semver.Version, logger log.Logger) error {
+	if ops.Type == "teams" && amVersion.LT(semver.MustParse("0.24.0")) {
+		return fmt.Errorf("'teams' set in 'opsgenieResponder' but supported in AlertManager >= 0.24.0 only")
+	}
 	return nil
 }
 
@@ -1474,6 +1538,23 @@ func (wcc *weChatConfig) sanitize(amVersion semver.Version, logger log.Logger) e
 
 func (sc *snsConfig) sanitize(amVersion semver.Version, logger log.Logger) error {
 	return sc.HTTPConfig.sanitize(amVersion, logger)
+}
+
+func (tc *telegramConfig) sanitize(amVersion semver.Version, logger log.Logger) error {
+	telegramAllowed := amVersion.GTE(semver.MustParse("0.24.0"))
+	if !telegramAllowed {
+		return fmt.Errorf(`invalid syntax in receivers config; telegram integration is available in Alertmanager >= 0.24.0`)
+	}
+
+	if tc.ChatID == 0 {
+		return errors.Errorf("mandatory field %q is empty", "chatID")
+	}
+
+	if tc.BotToken == "" {
+		return fmt.Errorf("mandatory field %q is empty", "botToken")
+	}
+
+	return tc.HTTPConfig.sanitize(amVersion, logger)
 }
 
 func (ir *inhibitRule) sanitize(amVersion semver.Version, logger log.Logger) error {