+++ pkg/client/versioned/clientset.go
@@ -22,6 +22,7 @@ import (
 
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned/typed/monitoring/v1"
 	monitoringv1alpha1 "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned/typed/monitoring/v1alpha1"
+	monitoringv1beta1 "github.com/prometheus-operator/prometheus-operator/pkg/client/versioned/typed/monitoring/v1beta1"
 	discovery "k8s.io/client-go/discovery"
 	rest "k8s.io/client-go/rest"
 	flowcontrol "k8s.io/client-go/util/flowcontrol"
@@ -31,6 +32,7 @@ type Interface interface {
 	Discovery() discovery.DiscoveryInterface
 	MonitoringV1() monitoringv1.MonitoringV1Interface
 	MonitoringV1alpha1() monitoringv1alpha1.MonitoringV1alpha1Interface
+	MonitoringV1beta1() monitoringv1beta1.MonitoringV1beta1Interface
 }
 
 // Clientset contains the clients for groups. Each group has exactly one
@@ -39,6 +41,7 @@ type Clientset struct {
 	*discovery.DiscoveryClient
 	monitoringV1       *monitoringv1.MonitoringV1Client
 	monitoringV1alpha1 *monitoringv1alpha1.MonitoringV1alpha1Client
+	monitoringV1beta1  *monitoringv1beta1.MonitoringV1beta1Client
 }
 
 // MonitoringV1 retrieves the MonitoringV1Client
@@ -49,6 +52,11 @@ func (c *Clientset) MonitoringV1() monitoringv1.MonitoringV1Interface {
 // MonitoringV1alpha1 retrieves the MonitoringV1alpha1Client
 func (c *Clientset) MonitoringV1alpha1() monitoringv1alpha1.MonitoringV1alpha1Interface {
 	return c.monitoringV1alpha1
+}
+
+// MonitoringV1beta1 retrieves the MonitoringV1beta1Client
+func (c *Clientset) MonitoringV1beta1() monitoringv1beta1.MonitoringV1beta1Interface {
+	return c.monitoringV1beta1
 }
 
 // Discovery retrieves the DiscoveryClient
@@ -99,6 +107,10 @@ func NewForConfigAndClient(c *rest.Config, httpClient *http.Client) (*Clientset,
 	if err != nil {
 		return nil, err
 	}
+	cs.monitoringV1beta1, err = monitoringv1beta1.NewForConfigAndClient(&configShallowCopy, httpClient)
+	if err != nil {
+		return nil, err
+	}
 
 	cs.DiscoveryClient, err = discovery.NewDiscoveryClientForConfigAndClient(&configShallowCopy, httpClient)
 	if err != nil {
@@ -122,6 +134,7 @@ func New(c rest.Interface) *Clientset {
 	var cs Clientset
 	cs.monitoringV1 = monitoringv1.New(c)
 	cs.monitoringV1alpha1 = monitoringv1alpha1.New(c)
+	cs.monitoringV1beta1 = monitoringv1beta1.New(c)
 
 	cs.DiscoveryClient = discovery.NewDiscoveryClient(c)
 	return &cs