+++ test/e2e/alertmanager_test.go
@@ -17,6 +17,7 @@ package e2e
 import (
 	"context"
 	"fmt"
+	"reflect"
 	"strconv"
 	"strings"
 	"testing"
@@ -38,6 +39,7 @@ import (
 	"github.com/prometheus-operator/prometheus-operator/pkg/alertmanager"
 	monitoringv1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	monitoringv1alpha1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1alpha1"
+	monitoringv1beta1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1beta1"
 )
 
 func testAMCreateDeleteCluster(t *testing.T) {
@@ -712,6 +714,82 @@ inhibit_rules:
 	}
 }
 
+func testAlertmanagerConfigVersions(t *testing.T) {
+	// Don't run Alertmanager tests in parallel. See
+	// https://github.com/prometheus/alertmanager/issues/1835 for details.
+	testCtx := framework.NewTestCtx(t)
+	defer testCtx.Cleanup(t)
+	ns := framework.CreateNamespace(context.Background(), t, testCtx)
+	framework.SetupPrometheusRBAC(context.Background(), t, testCtx, ns)
+
+	alertmanager := framework.MakeBasicAlertmanager("amconfig-versions", 1)
+	alertmanager.Spec.AlertmanagerConfigSelector = &metav1.LabelSelector{}
+	alertmanager, err := framework.CreateAlertmanagerAndWaitUntilReady(context.Background(), ns, alertmanager)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	amcfgV1alpha1 := &monitoringv1alpha1.AlertmanagerConfig{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "amcfg-v1alpha1",
+		},
+		Spec: monitoringv1alpha1.AlertmanagerConfigSpec{
+			Route: &monitoringv1alpha1.Route{
+				Receiver: "webhook",
+				Matchers: []monitoringv1alpha1.Matcher{{
+					Name:  "job",
+					Value: "webapp.+",
+					Regex: true,
+				}},
+			},
+			Receivers: []monitoringv1alpha1.Receiver{{
+				Name: "webhook",
+			}},
+		},
+	}
+
+	if _, err := framework.MonClientV1alpha1.AlertmanagerConfigs(alertmanager.Namespace).Create(context.Background(), amcfgV1alpha1, metav1.CreateOptions{}); err != nil {
+		t.Fatalf("failed to create v1alpha1 AlertmanagerConfig object: %v", err)
+	}
+
+	amcfgV1beta1Converted, err := framework.MonClientV1beta1.AlertmanagerConfigs(alertmanager.Namespace).Get(context.Background(), amcfgV1alpha1.Name, metav1.GetOptions{})
+	if err != nil {
+		t.Fatalf("failed to get v1beta1 AlertmanagerConfig object: %v", err)
+	}
+
+	expected := []monitoringv1beta1.Matcher{{Name: "job", Value: "webapp.+", MatchType: monitoringv1beta1.MatchRegexp}}
+	if !reflect.DeepEqual(amcfgV1beta1Converted.Spec.Route.Matchers, expected) {
+		t.Fatalf("expected %#v matcher, got %#v", expected, amcfgV1beta1Converted.Spec.Route.Matchers)
+	}
+
+	amcfgV1beta1 := &monitoringv1beta1.AlertmanagerConfig{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: "amcfg-v1beta1",
+		},
+		Spec: monitoringv1beta1.AlertmanagerConfigSpec{
+			Route: &monitoringv1beta1.Route{
+				Receiver: "webhook",
+				Matchers: []monitoringv1beta1.Matcher{{
+					Name:      "job",
+					Value:     "webapp.+",
+					MatchType: "=~",
+				}},
+			},
+			Receivers: []monitoringv1beta1.Receiver{{
+				Name: "webhook",
+			}},
+		},
+	}
+
+	if _, err := framework.MonClientV1beta1.AlertmanagerConfigs(alertmanager.Namespace).Create(context.Background(), amcfgV1beta1, metav1.CreateOptions{}); err != nil {
+		t.Fatalf("failed to create v1beta1 AlertmanagerConfig object: %v", err)
+	}
+
+	if _, err := framework.MonClientV1alpha1.AlertmanagerConfigs(alertmanager.Namespace).Get(context.Background(), amcfgV1beta1.Name, metav1.GetOptions{}); err != nil {
+		t.Fatalf("failed to get v1alpha1 AlertmanagerConfig object: %v", err)
+	}
+}
+
 func testAlertmanagerConfigCRD(t *testing.T) {
 	// Don't run Alertmanager tests in parallel. See
 	// https://github.com/prometheus/alertmanager/issues/1835 for details.
@@ -747,6 +825,21 @@ func testAlertmanagerConfigCRD(t *testing.T) {
 		},
 	}
 	if _, err := framework.KubeClient.CoreV1().Secrets(configNs).Create(context.Background(), testingKeySecret, metav1.CreateOptions{}); err != nil {
+		t.Fatal(err)
+	}
+
+	// telegram secret
+	telegramTestingSecret := "telegram-testing-secret"
+	telegramTestingbotTokenKey := "telegram-testing-bottoken-key"
+	telegramTestingKeySecret := &v1.Secret{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: telegramTestingSecret,
+		},
+		Data: map[string][]byte{
+			telegramTestingbotTokenKey: []byte("bipbop"),
+		},
+	}
+	if _, err := framework.KubeClient.CoreV1().Secrets(configNs).Create(context.Background(), telegramTestingKeySecret, metav1.CreateOptions{}); err != nil {
 		t.Fatal(err)
 	}
 
@@ -886,6 +979,17 @@ func testAlertmanagerConfigCRD(t *testing.T) {
 						Key: testingSecretKey,
 					},
 				}},
+				TelegramConfigs: []monitoringv1alpha1.TelegramConfig{{
+					APIURL: "https://telegram.api.url",
+					BotToken: &v1.SecretKeySelector{
+						LocalObjectReference: v1.LocalObjectReference{
+							Name: telegramTestingSecret,
+						},
+						Key: telegramTestingbotTokenKey,
+					},
+					ChatID: 12345,
+				}},
+
 				SNSConfigs: []monitoringv1alpha1.SNSConfig{
 					{
 						ApiURL: "https://sns.us-east-2.amazonaws.com",
@@ -1198,6 +1302,10 @@ receivers:
       access_key: 1234abc
       secret_key: 1234abc
     topic_arn: test-topicARN
+  telegram_configs:
+  - api_url: https://telegram.api.url
+    bot_token: bipbop
+    chat_id: 12345
 - name: %s/e2e-test-amconfig-sub-routes/e2e
   webhook_configs:
   - url: http://test.url